Refactor Azure module
