+++ tests/providers/azure/azure_test.py
@@ -26,7 +26,7 @@ from libcloudforensics.providers.azure import forensics
 # pylint: disable=line-too-long
 with mock.patch('libcloudforensics.providers.azure.internal.common.GetCredentials') as mock_creds:
   mock_creds.return_value = ('fake-subscription-id', mock.Mock())
-  with mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount._GetOrCreateResourceGroup') as mock_resource:
+  with mock.patch('libcloudforensics.providers.azure.internal.resource.AZResource.GetOrCreateResourceGroup') as mock_resource:
     # pylint: enable=line-too-long
     mock_resource.return_value = 'fake-resource-group'
     FAKE_ACCOUNT = account.AZAccount(
@@ -34,7 +34,7 @@ with mock.patch('libcloudforensics.providers.azure.internal.common.GetCredential
         default_region='fake-region'
     )
 
-FAKE_INSTANCE = compute.AZVirtualMachine(
+FAKE_INSTANCE = compute.AZComputeVirtualMachine(
     FAKE_ACCOUNT,
     '/a/b/c/fake-resource-group/fake-vm-name',
     'fake-vm-name',
@@ -42,7 +42,7 @@ FAKE_INSTANCE = compute.AZVirtualMachine(
     ['fake-zone']
 )
 
-FAKE_DISK = compute.AZDisk(
+FAKE_DISK = compute.AZComputeDisk(
     FAKE_ACCOUNT,
     '/a/b/c/fake-resource-group/fake-disk-name',
     'fake-disk-name',
@@ -50,7 +50,7 @@ FAKE_DISK = compute.AZDisk(
     ['fake-zone']
 )
 
-FAKE_BOOT_DISK = compute.AZDisk(
+FAKE_BOOT_DISK = compute.AZComputeDisk(
     FAKE_ACCOUNT,
     '/a/b/c/fake-resource-group/fake-boot-disk-name',
     'fake-boot-disk-name',
@@ -58,7 +58,7 @@ FAKE_BOOT_DISK = compute.AZDisk(
     ['fake-zone']
 )
 
-FAKE_SNAPSHOT = compute.AZSnapshot(
+FAKE_SNAPSHOT = compute.AZComputeSnapshot(
     FAKE_ACCOUNT,
     '/a/b/c/fake-resource-group/fake_snapshot_name',
     'fake_snapshot_name',
@@ -167,7 +167,7 @@ class TestAccount(unittest.TestCase):
 
     # If we don't specify a resource group name, the 'list_all' method should
     # be called.
-    instances = FAKE_ACCOUNT.ListInstances()
+    instances = FAKE_ACCOUNT.compute.ListInstances()
     mock_request.assert_called_with(mock.ANY, 'list_all')
     self.assertEqual(1, len(instances))
     self.assertIn('fake-vm-name', instances)
@@ -180,7 +180,7 @@ class TestAccount(unittest.TestCase):
     self.assertEqual(['fake-zone'], instance.zones)
 
     # If we specify a resource group name, the 'list' method should be called.
-    FAKE_ACCOUNT.ListInstances(
+    FAKE_ACCOUNT.compute.ListInstances(
         resource_group_name=instance.resource_group_name)
     mock_request.assert_called_with(
         mock.ANY, 'list', {'resource_group_name': instance.resource_group_name})
@@ -193,7 +193,7 @@ class TestAccount(unittest.TestCase):
 
     # If we don't specify a resource group name, the 'list' method should be
     # called.
-    disks = FAKE_ACCOUNT.ListDisks()
+    disks = FAKE_ACCOUNT.compute.ListDisks()
     mock_request.assert_called_with(mock.ANY, 'list')
     self.assertEqual(2, len(disks))
     self.assertIn('fake-disk-name', disks)
@@ -207,18 +207,18 @@ class TestAccount(unittest.TestCase):
 
     # If we specify a resource group name, the 'list_by_resource_group' method
     # should be called.
-    FAKE_ACCOUNT.ListDisks(resource_group_name=disk.resource_group_name)
+    FAKE_ACCOUNT.compute.ListDisks(resource_group_name=disk.resource_group_name)
     mock_request.assert_called_with(
         mock.ANY,
         'list_by_resource_group',
         {'resource_group_name': disk.resource_group_name})
 
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount.ListInstances')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZCompute.ListInstances')
   @typing.no_type_check
   def testGetInstance(self, mock_list_instances):
     """Test that a particular instance from an account is retrieved."""
     mock_list_instances.return_value = MOCK_LIST_INSTANCES
-    instance = FAKE_ACCOUNT.GetInstance('fake-vm-name')
+    instance = FAKE_ACCOUNT.compute.GetInstance('fake-vm-name')
     self.assertEqual('fake-vm-name', instance.name)
     self.assertEqual(
         '/a/b/c/fake-resource-group/fake-vm-name', instance.resource_id)
@@ -226,12 +226,12 @@ class TestAccount(unittest.TestCase):
     self.assertEqual('fake-region', instance.region)
     self.assertEqual(['fake-zone'], instance.zones)
 
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount.ListDisks')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZCompute.ListDisks')
   @typing.no_type_check
   def testGetDisk(self, mock_list_disks):
     """Test that a particular disk from an account is retrieved."""
     mock_list_disks.return_value = MOCK_LIST_DISKS
-    disk = FAKE_ACCOUNT.GetDisk('fake-disk-name')
+    disk = FAKE_ACCOUNT.compute.GetDisk('fake-disk-name')
     self.assertEqual('fake-disk-name', disk.name)
     self.assertEqual(
         '/a/b/c/fake-resource-group/fake-disk-name', disk.resource_id)
@@ -247,9 +247,9 @@ class TestAccount(unittest.TestCase):
     mock_create_disk.return_value.result.return_value = MOCK_DISK_COPY
     # CreateDiskFromSnapshot(
     #     snapshot=FAKE_SNAPSHOT, disk_name=None, disk_name_prefix='')
-    disk_from_snapshot = FAKE_ACCOUNT.CreateDiskFromSnapshot(
+    disk_from_snapshot = FAKE_ACCOUNT.compute.CreateDiskFromSnapshot(
         FAKE_SNAPSHOT)
-    self.assertIsInstance(disk_from_snapshot, compute.AZDisk)
+    self.assertIsInstance(disk_from_snapshot, compute.AZComputeDisk)
     self.assertEqual(
         'fake_snapshot_name_f4c186ac_copy', disk_from_snapshot.name)
     mock_create_disk.assert_called_with(
@@ -261,7 +261,7 @@ class TestAccount(unittest.TestCase):
     #     snapshot=FAKE_SNAPSHOT,
     #     disk_name='new-forensics-disk',
     #     disk_name_prefix='')
-    FAKE_ACCOUNT.CreateDiskFromSnapshot(
+    FAKE_ACCOUNT.compute.CreateDiskFromSnapshot(
         FAKE_SNAPSHOT, disk_name='new-forensics-disk')
     mock_create_disk.assert_called_with(
         FAKE_SNAPSHOT.resource_group_name,
@@ -270,7 +270,7 @@ class TestAccount(unittest.TestCase):
 
     # CreateDiskFromSnapshot(
     #     snapshot=FAKE_SNAPSHOT, disk_name=None, disk_name_prefix='prefix')
-    FAKE_ACCOUNT.CreateDiskFromSnapshot(
+    FAKE_ACCOUNT.compute.CreateDiskFromSnapshot(
         FAKE_SNAPSHOT, disk_name_prefix='prefix')
     mock_create_disk.assert_called_with(
         FAKE_SNAPSHOT.resource_group_name,
@@ -289,7 +289,7 @@ class TestAccount(unittest.TestCase):
             'name': 'StandardSSD_LRS'
         }
     }
-    FAKE_ACCOUNT.CreateDiskFromSnapshot(
+    FAKE_ACCOUNT.compute.CreateDiskFromSnapshot(
         FAKE_SNAPSHOT, disk_type='StandardSSD_LRS')
     mock_create_disk.assert_called_with(
         FAKE_SNAPSHOT.resource_group_name,
@@ -301,8 +301,8 @@ class TestAccount(unittest.TestCase):
   @mock.patch('azure.storage.blob._blob_service_client.BlobServiceClient.get_blob_client')
   @mock.patch('azure.storage.blob._blob_service_client.BlobServiceClient.get_container_client')
   @mock.patch('azure.storage.blob._blob_service_client.BlobServiceClient.__init__')
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount._DeleteStorageAccount')
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount._CreateStorageAccount')
+  @mock.patch('libcloudforensics.providers.azure.internal.storage.AZStorage.DeleteStorageAccount')
+  @mock.patch('libcloudforensics.providers.azure.internal.storage.AZStorage.CreateStorageAccount')
   @mock.patch('azure.mgmt.compute.v2020_05_01.operations._disks_operations.DisksOperations.create_or_update')
   @typing.no_type_check
   def testCreateDiskFromSnapshotUri(self,
@@ -326,13 +326,13 @@ class TestAccount(unittest.TestCase):
     blob_properties.return_value = mock.Mock(copy=mock.Mock(status='success'))
     mock_delete_storage_account.return_value = None
 
-    disk_from_snapshot_uri = FAKE_ACCOUNT.CreateDiskFromSnapshotURI(
+    disk_from_snapshot_uri = FAKE_ACCOUNT.compute.CreateDiskFromSnapshotURI(
         FAKE_SNAPSHOT, 'fake-snapshot-uri')
     #  hashlib.sha1('/a/b/c/fake-resource-group/fake_snapshot_name'.encode(
     #     'utf-8')).hexdigest()[:23] = bff00b08549ba8b975b2e70
     mock_create_storage_account.assert_called_with(
         'bff00b08549ba8b975b2e70', region='fake-region')
-    self.assertIsInstance(disk_from_snapshot_uri, compute.AZDisk)
+    self.assertIsInstance(disk_from_snapshot_uri, compute.AZComputeDisk)
     self.assertEqual(
         'fake_snapshot_name_f4c186ac_copy', disk_from_snapshot_uri.name)
     mock_create_disk.assert_called_with(
@@ -342,9 +342,9 @@ class TestAccount(unittest.TestCase):
 
   @mock.patch('sshpubkeys.SSHKey.parse')
   @mock.patch('libcloudforensics.scripts.utils.ReadStartupScript')
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount.GetInstance')
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount._GetInstanceType')
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount._CreateNetworkInterfaceForVM')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZCompute.GetInstance')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZCompute._GetInstanceType')
+  @mock.patch('libcloudforensics.providers.azure.internal.network.AZNetwork.CreateNetworkInterface')
   @mock.patch('azure.mgmt.compute.v2020_06_01.operations._virtual_machines_operations.VirtualMachinesOperations.create_or_update')
   @typing.no_type_check
   def testGetOrCreateAnalysisVm(self,
@@ -361,11 +361,11 @@ class TestAccount(unittest.TestCase):
     mock_script.return_value = ''
     mock_ssh_parse.return_value = None
 
-    vm, created = FAKE_ACCOUNT.GetOrCreateAnalysisVm(
+    vm, created = FAKE_ACCOUNT.compute.GetOrCreateAnalysisVm(
         FAKE_INSTANCE.name, 1, 4, 8192, '')
     mock_get_instance.assert_called_with(FAKE_INSTANCE.name)
     mock_vm.assert_not_called()
-    self.assertIsInstance(vm, compute.AZVirtualMachine)
+    self.assertIsInstance(vm, compute.AZComputeVirtualMachine)
     self.assertEqual('fake-vm-name', vm.name)
     self.assertFalse(created)
 
@@ -374,11 +374,11 @@ class TestAccount(unittest.TestCase):
     # created.
     mock_get_instance.side_effect = RuntimeError()
     mock_vm.return_value.result.return_value = MOCK_ANALYSIS_INSTANCE
-    vm, created = FAKE_ACCOUNT.GetOrCreateAnalysisVm(
+    vm, created = FAKE_ACCOUNT.compute.GetOrCreateAnalysisVm(
         'fake-analysis-vm-name', 1, 4, 8192, '')
     mock_get_instance.assert_called_with('fake-analysis-vm-name')
     mock_vm.assert_called()
-    self.assertIsInstance(vm, compute.AZVirtualMachine)
+    self.assertIsInstance(vm, compute.AZComputeVirtualMachine)
     self.assertEqual('fake-analysis-vm-name', vm.name)
     self.assertTrue(created)
 
@@ -387,23 +387,23 @@ class TestAccount(unittest.TestCase):
   def testListVMSizes(self, mock_list):
     """Test that instance types are correctly listed."""
     mock_list.return_value = MOCK_REQUEST_VM_SIZE
-    available_vms = FAKE_ACCOUNT.ListInstanceTypes()
+    available_vms = FAKE_ACCOUNT.compute.ListInstanceTypes()
     self.assertEqual(1, len(available_vms))
     self.assertEqual('fake-vm-type', available_vms[0]['Name'])
     self.assertEqual(4, available_vms[0]['CPU'])
     self.assertEqual(8192, available_vms[0]['Memory'])
 
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount.ListInstanceTypes')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZCompute.ListInstanceTypes')
   @typing.no_type_check
   def testGetInstanceType(self, mock_list_instance_types):
     """Test that the instance type given a configuration is correct."""
     # pylint: disable=protected-access
     mock_list_instance_types.return_value = MOCK_LIST_VM_SIZES
-    instance_type = FAKE_ACCOUNT._GetInstanceType(4, 8192)
+    instance_type = FAKE_ACCOUNT.compute._GetInstanceType(4, 8192)
     self.assertEqual('fake-vm-type', instance_type)
 
     with self.assertRaises(ValueError):
-      FAKE_ACCOUNT._GetInstanceType(666, 666)
+      FAKE_ACCOUNT.compute._GetInstanceType(666, 666)
     # pylint: enable=protected-access
 
   @mock.patch('azure.mgmt.resource.subscriptions.v2019_11_01.operations._subscriptions_operations.SubscriptionsOperations.list')
@@ -411,7 +411,7 @@ class TestAccount(unittest.TestCase):
   def testListSubscriptionIDs(self, mock_list):
     """Test that subscription IDs are correctly listed"""
     mock_list.return_value = MOCK_LIST_IDS
-    subscription_ids = FAKE_ACCOUNT.ListSubscriptionIDs()
+    subscription_ids = FAKE_ACCOUNT.resource.ListSubscriptionIDs()
     self.assertEqual(2, len(subscription_ids))
     self.assertEqual('fake-subscription-id-1', subscription_ids[0])
 
@@ -423,12 +423,13 @@ class TestAccount(unittest.TestCase):
     # pylint: disable=protected-access
     mock_create.return_value.result.return_value = MOCK_STORAGE_ACCOUNT
     mock_list_keys.return_value = MOCK_LIST_KEYS
-    account_id, account_key = FAKE_ACCOUNT._CreateStorageAccount('fakename')
+    account_id, account_key = FAKE_ACCOUNT.storage.CreateStorageAccount(
+        'fakename')
     self.assertEqual('fakestorageid', account_id)
     self.assertEqual('fake-key-value', account_key)
 
     with self.assertRaises(ValueError) as error:
-      _, _ = FAKE_ACCOUNT._CreateStorageAccount(
+      _, _ = FAKE_ACCOUNT.storage.CreateStorageAccount(
           'fake-non-conform-name')
     # pylint: enable=protected-access
     self.assertEqual('Storage account name fake-non-conform-name does not '
@@ -531,7 +532,7 @@ class TestAZVirtualMachine(unittest.TestCase):
   """Test Azure virtual machine class."""
   # pylint: disable=line-too-long
 
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount.ListDisks')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZCompute.ListDisks')
   @mock.patch('azure.mgmt.compute.v2020_06_01.operations._virtual_machines_operations.VirtualMachinesOperations.get')
   @typing.no_type_check
   def testGetBootDisk(self, mock_get_vm, mock_list_disk):
@@ -545,7 +546,7 @@ class TestAZVirtualMachine(unittest.TestCase):
         resource_group_name=FAKE_INSTANCE.resource_group_name)
     self.assertEqual('fake-boot-disk-name', boot_disk.name)
 
-  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZVirtualMachine.ListDisks')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZComputeVirtualMachine.ListDisks')
   @typing.no_type_check
   def testGetDisk(self, mock_list_disk):
     """Test that a particular disk from an instance is retrieved."""
@@ -559,7 +560,7 @@ class TestAZVirtualMachine(unittest.TestCase):
         'Disk non-existent-disk-name not found in instance: '
         '/a/b/c/fake-resource-group/fake-vm-name', str(error.exception))
 
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount.ListDisks')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZCompute.ListDisks')
   @mock.patch('azure.mgmt.compute.v2020_06_01.operations._virtual_machines_operations.VirtualMachinesOperations.get')
   @typing.no_type_check
   def testListDisks(self, mock_get_vm, mock_list_disk):
@@ -622,15 +623,15 @@ class TestForensics(unittest.TestCase):
   """Test Azure forensics file."""
   # pylint: disable=line-too-long, too-many-arguments
 
-  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZDisk.GetDiskType')
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount._GetOrCreateResourceGroup')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZComputeDisk.GetDiskType')
+  @mock.patch('libcloudforensics.providers.azure.internal.resource.AZResource.GetOrCreateResourceGroup')
   @mock.patch('libcloudforensics.providers.azure.internal.common.GetCredentials')
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount.ListSubscriptionIDs')
-  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZSnapshot.Delete')
-  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZDisk.Snapshot')
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount.GetDisk')
-  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZVirtualMachine.GetBootDisk')
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount.GetInstance')
+  @mock.patch('libcloudforensics.providers.azure.internal.resource.AZResource.ListSubscriptionIDs')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZComputeSnapshot.Delete')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZComputeDisk.Snapshot')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZCompute.GetDisk')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZComputeVirtualMachine.GetBootDisk')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZCompute.GetInstance')
   @mock.patch('azure.mgmt.compute.v2020_05_01.operations._disks_operations.DisksOperations.create_or_update')
   @typing.no_type_check
   def testCreateDiskCopy1(self,
@@ -668,18 +669,18 @@ class TestForensics(unittest.TestCase):
     mock_get_instance.assert_called_with('fake-vm-name')
     mock_get_boot_disk.assert_called_once()
     mock_get_disk.assert_not_called()
-    self.assertIsInstance(disk_copy, compute.AZDisk)
+    self.assertIsInstance(disk_copy, compute.AZComputeDisk)
     self.assertEqual('fake_snapshot_name_f4c186ac_copy', disk_copy.name)
 
-  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZDisk.GetDiskType')
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount._GetOrCreateResourceGroup')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZComputeDisk.GetDiskType')
+  @mock.patch('libcloudforensics.providers.azure.internal.resource.AZResource.GetOrCreateResourceGroup')
   @mock.patch('libcloudforensics.providers.azure.internal.common.GetCredentials')
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount.ListSubscriptionIDs')
-  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZSnapshot.Delete')
-  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZDisk.Snapshot')
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount.GetDisk')
-  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZVirtualMachine.GetBootDisk')
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount.GetInstance')
+  @mock.patch('libcloudforensics.providers.azure.internal.resource.AZResource.ListSubscriptionIDs')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZComputeSnapshot.Delete')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZComputeDisk.Snapshot')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZCompute.GetDisk')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZComputeVirtualMachine.GetBootDisk')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZCompute.GetInstance')
   @mock.patch('azure.mgmt.compute.v2020_05_01.operations._disks_operations.DisksOperations.create_or_update')
   @typing.no_type_check
   def testCreateDiskCopy2(self,
@@ -716,15 +717,15 @@ class TestForensics(unittest.TestCase):
     mock_get_boot_disk.assert_not_called()
     mock_get_disk.assert_called_once()
     mock_get_disk.assert_called_with('fake-disk-name')
-    self.assertIsInstance(disk_copy, compute.AZDisk)
+    self.assertIsInstance(disk_copy, compute.AZComputeDisk)
     self.assertEqual('fake_snapshot_name_f4c186ac_copy', disk_copy.name)
 
-  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZDisk.GetDiskType')
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount._GetOrCreateResourceGroup')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZComputeDisk.GetDiskType')
+  @mock.patch('libcloudforensics.providers.azure.internal.resource.AZResource.GetOrCreateResourceGroup')
   @mock.patch('libcloudforensics.providers.azure.internal.common.GetCredentials')
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount.ListSubscriptionIDs')
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount.ListDisks')
-  @mock.patch('libcloudforensics.providers.azure.internal.account.AZAccount.ListInstances')
+  @mock.patch('libcloudforensics.providers.azure.internal.resource.AZResource.ListSubscriptionIDs')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZCompute.ListDisks')
+  @mock.patch('libcloudforensics.providers.azure.internal.compute.AZCompute.ListInstances')
   @typing.no_type_check
   def testCreateDiskCopy3(self,
                           mock_list_instances,
